requirements.txt
asttokens==3.0.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.28.0
fsspec==2026.3.0
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
packaging==26.1
parso==0.8.6
pillow==12.2.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
pyzmq==27.1.0
setuptools==81.0.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.6.0
